Re-organize folder for footer pages.
